+++ tests/test_info.py
@@ -16,7 +16,7 @@ zogqb = xa.mfundinfo("001211", **ioconf)
 def test_fundreport():
     r = xa.FundReport("000827")
     assert r.get_report()[0][:2] == "广发"
-    assert r.analyse_report(1)["benchmark"][:3] == "本基金"
+    assert r.analyse_report(1)["bank"][:2] == "兴业"
     assert r.show_report_list(type_=0)[0]["FUNDCODE"] == "000827"
     assert r.get_report(id_="AN202003171376532533")[0][:2] == "广发"